frontend/requirements.txt
flask
requests
python-dotenv
ms_identity_python[flask] @ git+https://github.com/azure-samples/ms-identity-python@0.9
langchain[openai]
langgraph